+++ lib/src/theme.dart
@@ -4,13 +4,35 @@ import 'package:flutter/widgets.dart';
 import 'inherit.dart';
 import 'wrap.dart';
 
-mixin Theme {
-  /// Default background of the theme.
-  /// It's recommended to use a final parameter rather than a getter.
+/// All handled theme data should have basic properties defined in this mixin.
+///
+/// The [foreground] and [background] are elementary components of a theme,
+/// without such parameters, the theme will be useless.
+/// So that they are required.
+///
+/// It's recommended to use a final parameter rather than a getter.
+/// It is defined as a getter here because of the limit of Dart syntax.
+/// For example:
+///
+/// ```dart
+/// class Theme with ThemeMixin {
+///   const Theme({
+///     required this.background,
+///     required this.foreground,
+///   })
+///
+///   @override
+///   final Color background;
+///
+///   @override
+///   final Color foreground;
+/// }
+/// ```
+mixin ThemeMixin {
+  /// Default background of the theme, see [ThemeMixin].
   Color get background;
 
-  /// Default foreground, text and icon color, of the theme.
-  /// It's recommended to use a final parameter rather than a getter.
+  /// Default foreground, text and icon color, of the theme, see [ThemeMixin].
   Color get foreground;
 }
 
@@ -27,7 +49,7 @@ extension WrapTheme on Widget {
   /// [light] theme directly when there's no [dark] theme,
   /// because even if the [dark] theme is not provided,
   /// it might be modifier by its descendants.
-  Widget theme<T extends Theme>({
+  Widget theme<T extends ThemeMixin>({
     required T light,
     T? dark,
     ThemeMode mode = ThemeMode.system,
@@ -40,11 +62,11 @@ extension WrapTheme on Widget {
       );
 }
 
-/// Handle a [Theme] and the [ThemeMode],
+/// Handle a [ThemeMixin] and the [ThemeMode],
 /// and it will also provide the [Brightness] of current theme.
-/// And you can also modify current [Theme] and [ThemeMode] from
+/// And you can also modify current [ThemeMixin] and [ThemeMode] from
 /// its descendants in the widget tree via the context.
-class ThemeHandler<T extends Theme> extends StatefulWidget {
+class ThemeHandler<T extends ThemeMixin> extends StatefulWidget {
   const ThemeHandler({
     super.key,
     required this.light,
@@ -62,7 +84,7 @@ class ThemeHandler<T extends Theme> extends StatefulWidget {
   State<ThemeHandler<T>> createState() => _ThemeHandlerState<T>();
 }
 
-class _ThemeHandlerState<T extends Theme> extends State<ThemeHandler<T>> {
+class _ThemeHandlerState<T extends ThemeMixin> extends State<ThemeHandler<T>> {
   late T _light = widget.light;
   late T _dark = widget.dark;
   late ThemeMode _mode = widget.mode;